monday 01-27
Changed the name of skeleton.c to phase1.c. I also added memset to initialize ProcTable. Unsure if that is the correct way to initialize it. I also added a for loop to Fork1() which will look for a PID on the ProcTable that is 0 which inicates that array member is empty, and as such a new process can be placed there.